mejorando

+++ Dominio/Articulo.cs
@@ -11,10 +11,12 @@ namespace Dominio
         public string codigo { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
+
         public Marca marca { get; set; }
+
         public Categoria categoria { get; set; }
         public string UrlImagen { get; set; }
-        public float Precio { get; set; }
+        public decimal Precio { get; set; }
 
     }
 }
